requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pytest==7.4.0
jupyter==1.0.0